GSH_flask_app/requirements.txt
pyserial
time
requests
json
flask
jinja2